print json error

+++ vamtb/vamtb.py
@@ -365,10 +365,13 @@ def dbscan(ctx):
             try:
                 if var.store_update(confirm=False if ctx.obj['force'] else True):
                     stored += 1
-            except (VarMalformed, VarMetaJson) as e:
-                error(f"Var {var.var} malformed.")
-            except:
-                pass
+            except VarMalformed as e:
+                error(f"Var {var.var} malformed")
+            except NoMetaJson:
+                error(f"Var {var.var} malformed (no meta found)")
+            except VarMetaJson as e:
+                error(f"Var {var.var} has something wrong in meta: {e}")
+
     info(f"{stored} var files stored")
 
 
@@ -385,14 +388,11 @@ def dbclean(ctx):
 
     files = search_files_indir(ctx.obj['dir'], f".*\.var", ign=True)
     varnames = set([ e.with_suffix("").name for e in files ])
-    meh = [e for e in varnames if "NoOC" in e]
     dbvars = set(Dbs.get_vars())
-    meh2 = [e for e in dbvars if "NoOC" in e]
-    print("\n".join(sorted(meh)))
-    print("\n".join(sorted(meh2)))
-    exit(0)
     diff = sorted(list(varnames - dbvars) + list(dbvars - varnames))
     for var in diff:
+        if var not in dbvars:
+            continue
         print(f"Var {red(var)} is in DB but not on disk  ")
         if ctx.obj['force'] or input("Delete from DB: Y [N] ?").upper() == "Y":
             varfile = VarFile(var, use_db=True)
